Block high-severity Solidity regressions before deployment
Use Foundry stable lint as a fail-closed production-source gate in local checks and CI, and make venue-mask shifts explicitly uint256-sized.

Constraint: Add no new static-analysis dependency and keep test-fixture warning cleanup separate.
Rejected: Denying every existing medium and low warning immediately | would mix fixture cleanup with the production security gate.
Confidence: high
Scope-risk: narrow
Directive: Keep high-severity production lint warning-free; expand warning budgets in a dedicated follow-up.
Tested: forge lint --severity high --deny warnings src; forge test --offline (582 passed); scripts/check.sh; git diff --cached --check
Not-tested: live Anvil E2E because runtime behavior is unchanged

+++ src/execution/VenueSelector.sol
@@ -16,7 +16,7 @@ import {ComplianceDecision, VenueType} from "../types/ComplianceTypes.sol";
 ///        - otherwise   => must equal keccak256(abi.encode(venue)), binding this exact venue.
 contract VenueSelector is IVenueSelector {
     function validate(address venue, VenueType vtype, ComplianceDecision calldata d) external pure returns (bool) {
-        return d.allowed && (d.allowedVenueTypes & (1 << uint256(vtype))) != 0
+        return d.allowed && (d.allowedVenueTypes & (uint256(1) << uint256(vtype))) != 0
             && (d.allowedVenuesHash == bytes32(0) || d.allowedVenuesHash == keccak256(abi.encode(venue)));
     }
 }
